+++ joulescope_ui/app.py
@@ -190,6 +190,12 @@ SETTINGS = {
         ],
         'default': 'desktop',
     },
+    'intel_graphics_prompt': {
+        'dtype': 'bool',
+        'brief': N_('Prompt to switch from Intel graphics to software renderer.'),
+        'default': True,
+        'flags': ['skip_undo'],
+    },
 }